Agent: Rename send_all_events_to_island.py -> agent_event_forwarder.py
